fix: plugin->cog change
